+++ test/index.test.js
@@ -1,93 +1,123 @@
-const Blackrik = require('../src/index');
-const Application = require('../src/core/Blackrik');
-const exampleInstance = require('./_resources/testApp/config');
-
-const defaultAdapters = Object.freeze(require('../src/adapters'));
-const httpMethods = Object.freeze(require('../src/resources/httpMethods'));
-const Errors = Object.freeze(require('../src/core/Errors'));
-
 jest.mock('../src/core/Blackrik', () => {
+    const init = jest.fn();
     const executeCommand = jest.fn();
     const scheduleCommand = jest.fn();
     const executeQuery = jest.fn();
+    const deleteAggregate = jest.fn();
     const start = jest.fn();
     const stop = jest.fn();
     return jest.fn(() => {
         return {
-            init: jest.fn(() => 'Test string'),
+            init,
             executeCommand,
             scheduleCommand,
             executeQuery,
+            deleteAggregate,
             start,
             stop
         };
     });
 });
 
-let app;
-let testObj;
+const Blackrik = require('../src/index');
+const _Blackrik = require('../src/core/Blackrik');
+const exampleConfig = require('./_resources/testApp/config');
+
+const defaultAdapters = require('../src/adapters');
+const httpMethods = require('../src/resources/httpMethods');
+const Errors = require('../src/core/Errors');
+
+let instance;
+let blackrik;
 beforeEach(() => {
-    app = new Blackrik(exampleInstance);
-    testObj = new Application(app);
+    jest.clearAllMocks();
+    instance = new Blackrik(exampleConfig);
+    blackrik = new _Blackrik(instance);
 });
-describe('Test static metods', () => {
-    test('ADAPTERS', () => {
-        const result = Blackrik.ADAPTERS;
-        expect(result).toBe(defaultAdapters);
+describe('Blackrik', () => {
+    describe('has static properties', () => {
+        test('ADAPTERS', () => {
+            expect(Blackrik.ADAPTERS).toBe(defaultAdapters);
+        });
+        test('HTTP_METHODS', () => {
+            expect(Blackrik.HTTP_METHODS).toBe(httpMethods);
+        });
+        test('ERRORS', () => {
+            expect(Blackrik.ERRORS).toBe(Errors);
+        });
     });
-    test('HTTP_METHODS', () => {
-        const result = Blackrik.HTTP_METHODS;
-        expect(result).toBe(httpMethods);
-    });
-    test('ERRORS', () => {
-        const result = Blackrik.ERRORS;
-        expect(result).toBe(Errors);
-    });
-});
 
-describe('Test commands / query', () => {
-    test('executeCommand', async () => {
-        const testCommand = 'test';
-        await app.executeCommand(testCommand);
+    describe('encapsules functions', () => {
+        test('executeCommand', async () => {
+            const params = [{}, 'invalid'];
+            const expectedParams = params.slice(0, 1);
 
-        expect(testObj.executeCommand).toHaveBeenCalledWith(testCommand);
-    });
-    test('scheduleCommand', async () => {
-        const timestamp = 4224;
-        const testCommand = 'test';
+            await instance.executeCommand(...params);
 
-        await app.scheduleCommand(timestamp, testCommand);
-        expect(testObj.scheduleCommand).toHaveBeenCalledWith(timestamp, testCommand);
-    });
-    test('executeQuery', async () => {
-        app.executeQuery();
-    
-        expect(app.executeQuery).toHaveBeenCalled();
-    });
-});
+            expect(blackrik.executeCommand).toHaveBeenCalledWith(...expectedParams);
+        });
+        test('scheduleCommand', async () => {
+            const params = [0, {}, 'invalid'];
+            const expectedParams = params.slice(0, 2);
 
-describe('Test start', () => {
-    test('Start sucessfull - process not yet started', async () => {
-        await app.start();
+            await instance.scheduleCommand(...params);
+
+            expect(blackrik.scheduleCommand).toHaveBeenCalledWith(...expectedParams);
+        });
+        test('executeQuery', async () => {
+            const params = ['', '', {}, 'invalid'];
+            const expectedParams = params.slice(0, 3);
+
+            await instance.executeQuery(...params);
         
-        expect(testObj.start).toHaveBeenCalled();
-    });
-    test('Start sucessfull - already started', async () => {
-        await app.start();
-        await app.start();
-    });
-});
+            expect(blackrik.executeQuery).toHaveBeenCalledWith(...expectedParams);
+        });
+        describe('deleteAggregate', () => {
+            test('encapsuled', async () => {
+                const params = ['', '', {}, 'invalid'];
+                const expectedParams = params.slice(0, 3);
 
-describe('Test stop', () => {
-    test('stop without running application', async () => {
-        await app.stop();
+                await instance.deleteAggregate(...params);
+
+                expect(blackrik.deleteAggregate).toHaveBeenCalledWith(...expectedParams);
+            });
+            test('optional argument', async () => {
+                const params = ['', ''];
+                const expectedParams = params.concat(null);
+
+                await instance.deleteAggregate(...params);
+                
+                expect(blackrik.deleteAggregate).toHaveBeenCalledWith(...expectedParams);
+            });
+        });
+
+        describe('start', () => {
+            test('initial start', async () => {
+                await instance.start();
+                
+                expect(blackrik.start).toHaveBeenCalledTimes(1);
+            });
+            test('prevents double execution if already started', async () => {
+                await instance.start();
+                await instance.start();
+
+                expect(blackrik.start).toHaveBeenCalledTimes(1);
+            });
+        });
+
+        describe('stop', () => {
+            test('stop while not running', async () => {
+                await instance.stop();
+                
+                expect(blackrik.stop).not.toHaveBeenCalled();
+            });
+            test('stop while running', async () => {
+                await instance.start();
+                await instance.stop();
+                await instance.stop();
         
-        expect(testObj.stop).not.toHaveBeenCalled();
-    });
-    test('stop a running application', async () => {
-        await app.start();
-        await app.stop();
-
-        expect(testObj.stop).toHaveBeenCalled();
+                expect(blackrik.stop).toHaveBeenCalledTimes(1);
+            });
+        });
     });
 });